Start refactoring default special variables out
Remove the &key from '(defun sparql-server-put-data-request) and '(defun
sparql-server-status-request), and ensure the 'server-url parameter is
first in sequence.

I intend for this code to be used primarily as a library on which
applications are built, so I'm engineering out anything oriented around
single-person interaction; it's still easy enough to #'defun a suitable
dynamic variable when working interactively.

Moving the server-url parameter first may be premature optimisation, but
it's consistent with the way I'm doing things elsewhere, and also makes
it easier to turn it into a method if I decide to go that way.