Auto-quote plaintext objects when inserting triples
The object can be either an XML-formatted URI like the subject and the
predicate, or plaintext. The catch is that plaintext needs to be wrapped
in quotes.

Thus, add a couple of helper functions ('angle-bracket-delimited-p and
'quote-plaintext) for 'insert-triple to use.

Also update 'test-insert-triple to try the plaintext side of this.