gh/1: Add 'add-triples and remove 'add-triple
'add-triples makes most sense, as I expect it to be the rule that
multiple triples will be inserted in a batch, and that it will be an
exception to insert just one triple at a time.

Remove 'add triple, as there's little point having both of these.

Update the test file to account for the change.